Create sumOfArrayMultiThread.c
This code creates two threads and computes the sum of an array. It divides the array into two parts, the sum of the first half is computed by thread 0 and the sum of the second half is computed by thread 1. Both treads run concurrently. 

This code also computes the execution time of the code.